# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type StudentRole = typeof StudentRole[keyof typeof StudentRole];


export const StudentRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface Student {
  id: number;
  userId: string;
  role: StudentRole;
  name: string;
  email: string;
  /** @nullable */
  age: number | null;
  /** @nullable */
  profileImageObjectPath: string | null;
  /** @nullable */
  goal: string | null;
  /** @nullable */
  posturalAnalysis: string | null;
  /** @nullable */
  priorExperience: string | null;
  /** @nullable */
  phone: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface StudentUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  profileImageObjectPath?: string | null;
  /** @nullable */
  goal?: string | null;
  /** @nullable */
  posturalAnalysis?: string | null;
  /** @nullable */
  priorExperience?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface ClassSession {
  id: number;
  title: string;
  instructor: string;
  /**
     * 0 = Sunday .. 6 = Saturday
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  /** 24h time, e.g. 18:00 */
  startTime: string;
  /** 24h time, e.g. 19:00 */
  endTime: string;
  level: string;
  capacity: number;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

export interface ClassInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  instructor: string;
  /**
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  /** @minLength 1 */
  startTime: string;
  /** @minLength 1 */
  endTime: string;
  /** @minLength 1 */
  level: string;
  /** @minimum 1 */
  capacity: number;
  description?: string;
}

export interface ClassUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  instructor?: string;
  /**
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek?: number;
  /** @minLength 1 */
  startTime?: string;
  /** @minLength 1 */
  endTime?: string;
  /** @minLength 1 */
  level?: string;
  /** @minimum 1 */
  capacity?: number;
  /** @nullable */
  description?: string | null;
}

export interface Package {
  id: number;
  name: string;
  description: string;
  /** e.g. Pole, Aerial, Dance, Open Studio */
  category: string;
  price: number;
  /**
     * Set when isPromo is true to show a struck-through price.
     * @nullable
     */
  originalPrice: number | null;
  sessionCount: number;
  validityDays: number;
  /** @nullable */
  imageObjectPath: string | null;
  isPromo: boolean;
  active: boolean;
  createdAt: string;
}

export interface PackageInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  originalPrice?: number;
  /** @minimum 1 */
  sessionCount: number;
  /** @minimum 1 */
  validityDays: number;
  imageObjectPath?: string;
  isPromo?: boolean;
  active?: boolean;
}

export interface PackageUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @minimum 1 */
  sessionCount?: number;
  /** @minimum 1 */
  validityDays?: number;
  /** @nullable */
  imageObjectPath?: string | null;
  isPromo?: boolean;
  active?: boolean;
}

export interface PaymentMethod {
  id: number;
  /** e.g. GCash, Bank Transfer */
  label: string;
  qrImageObjectPath: string;
  /** @nullable */
  instructions: string | null;
  active: boolean;
  createdAt: string;
}

export interface PaymentMethodInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  qrImageObjectPath: string;
  instructions?: string;
  active?: boolean;
}

export interface PaymentMethodUpdate {
  /** @minLength 1 */
  label?: string;
  /** @minLength 1 */
  qrImageObjectPath?: string;
  /** @nullable */
  instructions?: string | null;
  active?: boolean;
}

export type PurchaseStatus = typeof PurchaseStatus[keyof typeof PurchaseStatus];


export const PurchaseStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
} as const;

export interface Purchase {
  id: number;
  studentId: number;
  packageId: number;
  packageName: string;
  amount: number;
  screenshotObjectPath: string;
  status: PurchaseStatus;
  /** @nullable */
  reviewNotes: string | null;
  submittedAt: string;
  /** @nullable */
  reviewedAt: string | null;
}

export interface PurchaseInput {
  packageId: number;
  /** @minimum 0 */
  amount: number;
  /** @minLength 1 */
  screenshotObjectPath: string;
}

export interface PurchaseReview {
  status: PurchaseStatus;
  reviewNotes?: string;
}

export type RentalStatus = typeof RentalStatus[keyof typeof RentalStatus];


export const RentalStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
} as const;

export interface Rental {
  id: number;
  requesterName: string;
  email: string;
  phone: string;
  /** Calendar date, YYYY-MM-DD */
  date: string;
  startTime: string;
  endTime: string;
  purpose: string;
  status: RentalStatus;
  /** @nullable */
  reviewNotes: string | null;
  createdAt: string;
}

export interface RentalInput {
  /** @minLength 1 */
  requesterName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  startTime: string;
  /** @minLength 1 */
  endTime: string;
  /** @minLength 1 */
  purpose: string;
}

export interface RentalReview {
  status: RentalStatus;
  reviewNotes?: string;
}

export interface DashboardSummary {
  totalStudents: number;
  pendingPurchases: number;
  pendingRentals: number;
  confirmedRevenue: number;
  upcomingClassCount: number;
}

export type ListPackagesParams = {
activeOnly?: boolean;
};

export type ListPurchasesParams = {
status?: PurchaseStatus;
};

export type ListRentalsParams = {
status?: RentalStatus;
};