# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Lourde Fitness Studio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Students ----------
  /students/me:
    get:
      tags: [Students]
      operationId: getMyProfile
      summary: Get or JIT-create the current signed-in student's profile
      responses:
        "200":
          description: Current student profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Students]
      operationId: updateMyProfile
      summary: Update the current signed-in student's own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /students:
    get:
      tags: [Students]
      operationId: listStudents
      summary: List all students (admin only)
      responses:
        "200":
          description: List of students.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /students/{id}:
    get:
      tags: [Students]
      operationId: getStudent
      summary: Get a single student by id (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Student not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Students]
      operationId: updateStudent
      summary: Update a student record (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Student not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Classes / Timetable ----------
...
  /rentals:
    get:
      tags: [Rentals]
      operationId: listRentals
      summary: List all studio rental requests (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RentalStatus"
      responses:
        "200":
          description: List of rental requests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Rentals]
      operationId: createRental
      summary: Submit a studio rental reservation request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalInput"
      responses:
        "201":
          description: Created rental request, pending review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rentals/{id}:
    patch:
      tags: [Rentals]
      operationId: reviewRental
      summary: Confirm or reject a studio rental request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalReview"
      responses:
        "200":
          description: Updated rental.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Rental not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Dashboard ----------
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---------- Storage ----------
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    # ---------- Students ----------
    StudentRole:
      type: string
      enum: [student, admin]
    Student:
      type: object
      required:
        [
          id,
          userId,
          role,
          name,
          email,
          age,
          profileImageObjectPath,
          goal,
          posturalAnalysis,
          priorExperience,
          phone,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          $ref: "#/components/schemas/StudentRole"
        name:
          type: string
        email:
          type: string
        age:
          type: ["integer", "null"]
        profileImageObjectPath:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        posturalAnalysis:
          type: ["string", "null"]
        priorExperience:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: ["integer", "null"]
        profileImageObjectPath:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        posturalAnalysis:
          type: ["string", "null"]
        priorExperience:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ---------- Classes ----------
...
    RentalStatus:
      type: string
      enum: [pending, confirmed, rejected]
    Rental:
      type: object
      required:
        [
          id,
          requesterName,
          email,
          phone,
          date,
          startTime,
          endTime,
          purpose,
          status,
          reviewNotes,
          createdAt,
        ]
      properties:
        id:
          type: integer
        requesterName:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
          description: "Calendar date, YYYY-MM-DD"
        startTime:
          type: string
        endTime:
          type: string
        purpose:
          type: string
        status:
          $ref: "#/components/schemas/RentalStatus"
        reviewNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RentalInput:
      type: object
      required: [requesterName, email, phone, date, startTime, endTime, purpose]
      properties:
        requesterName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        date:
          type: string
          minLength: 1
        startTime:
          type: string
          minLength: 1
        endTime:
          type: string
          minLength: 1
        purpose:
          type: string
          minLength: 1
    RentalReview:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/RentalStatus"
        reviewNotes:
          type: string

    # ---------- Dashboard ----------
...