# lib/api-zod/src/generated/types/rental.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */
import type { RentalStatus } from './rentalStatus';

export interface Rental {
  id: number;
  requesterName: string;
  email: string;
  phone: string;
  /** Calendar date, YYYY-MM-DD */
  date: string;
  startTime: string;
  endTime: string;
  purpose: string;
  status: RentalStatus;
  /** @nullable */
  reviewNotes: string | null;
  createdAt: Date;
}