# lib/api-zod/src/generated/types/rentalInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */

export interface RentalInput {
  /** @minLength 1 */
  requesterName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  startTime: string;
  /** @minLength 1 */
  endTime: string;
  /** @minLength 1 */
  purpose: string;
}