# lib/api-zod/src/generated/types/rentalStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */

export type RentalStatus = typeof RentalStatus[keyof typeof RentalStatus];


export const RentalStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
} as const;