# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */
import type { StudentRole } from './studentRole';

export interface Student {
  id: number;
  userId: string;
  role: StudentRole;
  name: string;
  email: string;
  /** @nullable */
  age: number | null;
  /** @nullable */
  profileImageObjectPath: string | null;
  /** @nullable */
  goal: string | null;
  /** @nullable */
  posturalAnalysis: string | null;
  /** @nullable */
  priorExperience: string | null;
  /** @nullable */
  phone: string | null;
  createdAt: Date;
  updatedAt: Date;
}